make the router ignore command line options
There is a small caveat in that we don't know which
options are switches until we know which action we're
planning on invoking.

I'm not sure what the eventual answer is... perhaps let
the action specified in the pattern participate in the
matching. In fact, this might be possible since we do
know ahead of time which action is associated with each
rule